requirements.txt
Faker
python-dotenv